pyproject.toml
[project]
name = "ada-pulse-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=3.10.1",
    "pandas>=3.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
